+++ R/pas_load.R
@@ -168,8 +168,17 @@ pas_load <- function(
   # NOTE:  loading might fail.
   
   if ( !successful ) {
-    stop(paste("Data file could not be loaded after ", retries, " tries"),
-         call. = FALSE)
+    if ( !is.null(getArchiveBaseDir()) ) {
+      stop(
+        sprintf("The requested 'pas' object is not found in %s", getArchiveBaseDir()),
+        "\n\nYou can unset the data archive BASE_DIR with:\n\n",
+        "  removeArchiveBaseDir()",
+        call. = FALSE
+      )
+    } else {
+      stop(sprintf("Data file could not be loaded after %d tries", retries),
+           call. = FALSE)
+    }
   }